ML-8 (#1)
Bug fixes and tests
