+++ src/main/app/business/util/ml_logger/logger.py
@@ -3,6 +3,14 @@ import logging
 _log_format = f"%(asctime)s - [%(levelname)s] - %(name)s - (%(filename)s).%(funcName)s(%(lineno)d) - %(message)s"
 
 
+def get_logger(name):
+    logger = logging.getLogger(name)
+    logger.setLevel(logging.INFO)
+    # logger.addHandler(_get_file_handler())
+    logger.addHandler(_get_stream_handler())
+    return logger
+
+
 def _get_file_handler():
     file_handler = logging.FileHandler("x.log")
     file_handler.setLevel(logging.WARNING)
@@ -15,11 +23,3 @@ def _get_stream_handler():
     stream_handler.setLevel(logging.INFO)
     stream_handler.setFormatter(logging.Formatter(_log_format))
     return stream_handler
-
-
-def get_logger(name):
-    logger = logging.getLogger(name)
-    logger.setLevel(logging.INFO)
-    # logger.addHandler(_get_file_handler())
-    logger.addHandler(_get_stream_handler())
-    return logger
