+++ src/test/app/business/util/logger_test.py
@@ -1,0 +1,33 @@
+import logging
+
+import business.util.ml_logger.logger as log_module
+
+
+class LoggerTests:
+    def setup(self):
+        print("basic setup into class")
+
+    def teardown(self):
+        print("basic teardown into class")
+
+    @classmethod
+    def setup_class(cls):
+        print("class setup")
+
+    @classmethod
+    def teardown_class(cls):
+        print("class teardown")
+
+    def setup_method(self, method):
+        print("method setup")
+
+    def teardown_method(self, method):
+        print("method teardown")
+
+    def test_created_logger(self):
+        logger_name = "TEST_LOGGER"
+        log = log_module.get_logger(logger_name)
+
+        assert (log is not None)
+        assert (log.name == logger_name)
+        assert (log.level == logging.INFO)
